+++ icu4c/source/common/uelement.h
@@ -56,7 +56,7 @@ typedef UBool U_CALLCONV UElementsAreEqual(const UElement e1, const UElement e2)
  * @param e2 An element (object or integer)
  * @return 0 if the two elements are equal, -1 if e1 is < e2, or +1 if e1 is > e2.
  */
-typedef int8_t U_CALLCONV UElementComparator(UElement e1, UElement e2);
+typedef int32_t U_CALLCONV UElementComparator(UElement e1, UElement e2);
 
 /**
  * An element assignment function.  It may copy an integer, copy
